tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-direction-wheel-scroll.html

<!DOCTYPE html>
<title>@container: scroll-state(direction) mouse wheel scroll</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#direction">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="/dom/events/scrolling/scroll_support.js"></script>
<script src="/css/css-scroll-snap/support/common.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style>
  #scroller {
    width: 200px;
    height: 200px;
    container-type: scroll-state;
    overflow-y: scroll;
  }
  #filler {
    height: 600px;
  }
  #target {
    --y: no;
    @container scroll-state(direction: y) {
      --y: yes;
    }
  }
</style>
<div id="scroller">
  <div id="filler">
    <div id="target"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    let scrollEndPromise = waitForScrollEndFallbackToDelayWithoutScrollEvent(scroller);
    await new test_driver.Actions()
          .scroll(0, 0, 0, 100, { origin: scroller, duration: 100 })
          .send();
    await scrollEndPromise;
    await waitForAnimationFrames(2);
    assert_equals(scroller.scrollTop, 100);
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
  }, "Mouse wheel scroll");

</script>
